etapas 3/4

+++ app/Controllers/Usuarios.php
@@ -3,9 +3,10 @@ namespace App\Controllers;
 
 use App\Models\UsuariosModel;
 use CodeIgniter\API\ResponseTrait;
-use CodeIgniter\Controller;
+use DateInterval;
+use DateTime;
 
-class Usuarios extends Controller
+class Usuarios extends BaseController
 {
   use ResponseTrait;
 
@@ -20,12 +21,26 @@ class Usuarios extends Controller
 
     $data = [
       "id" => \App\Libraries\UUID::v4(),
-      "nome" => $data->nome,
-      "email" => $data->email,
+      "nome" => @$data->nome,
+      "email" => @$data->email,
       "data_nascimento" => @$data->data_nascimento,
+      "saldo" => (isset($data->saldo))? $data->saldo: 0,
       "created_at" => date('Y-m-d H:i:s'),
       "updated_at" => "null",
     ];
+
+    if (!isset($data['data_nascimento'])) {
+      return $this->fail(['message'=>'Campo obrigatório: Data de Nascimento'], 204);
+    }
+    // validando idade
+    $hoje= new DateTime(date('Y-m-d'));
+    $niver= new DateTime($data['data_nascimento']);
+    $idade= $niver->diff($hoje);
+
+    if ($idade->y < 18) {
+      return $this->fail(['message'=>'Usuário deve ter 18 anos ou mais para criar conta']);
+    }
+    echo '<pre>'; var_dump($idade->y); die;
 
     if (!(!!$model->insert($data, false))) {
       return $this->fail($model->validation->getErrors());
@@ -75,12 +90,46 @@ class Usuarios extends Controller
     $params = $this->request->uri->getSegment(2);
     $model= new UsuariosModel();
 
-    $model->delete($params, true);
+    // verificando se o usuário tem saldo e/ou movimentação
+    $movi= new \App\Controllers\Movimentacoes();
+    $list= $movi->getList($params, true);
 
-    if ( is_null($model->find($params)) ) {
+    if (empty($list['usuario'])) {
       return $this->respondNoContent('Usuário não encontrado');
     }
 
+    if ($list['usuario']['saldo'] != 0 or !empty($list['movimentacoes'])) {
+      return $this->fail('Usuário já possui saldo e/ou movimentação cadastrado');
+    }
+
+    $model->delete($params, true);
+
+    if ( is_null($model->find($params)) ) {
+      return $this->respond(['message'=>'Usuário removido com sucesso']);
+    }
+
     return $this->respondDeleted($params);
+  }
+
+  /**
+   * altera o saldo de um usuário
+   */
+  public function addSaldo()
+  {
+    $data= $this->request->getJSON();
+    $token= $this->getToken();
+
+    $model = new UsuariosModel();
+
+    $data = [
+      "saldo" => $data->saldo,
+      "updated_at" => date('Y-m-d H:i:s'),
+    ];
+
+    if (!(!!$model->update($token, $data, false))) {
+      return $this->fail($model->validation->getErrors());
+    }
+
+    return $this->respond($data);
   }
 }
